A lot of stuff, textures

+++ src/render/vertex.hpp
@@ -3,8 +3,6 @@
 #include <glm/detail/type_vec3.hpp>
 #include <glm/detail/type_vec4.hpp>
 #include <glm/detail/type_vec.hpp>
-//
-#include <alias/int.hpp>
 
 namespace render
 {
@@ -13,7 +11,7 @@ namespace render
 struct Vertex
 {
     glm::vec3  pos;
-    glm::uvec2 tex_pos;
+    glm::vec2 tex_pos;
     glm::vec4  color;
 };
 #pragma pack(pop)
